Got a very simple job creation page going. Can select solvers and benchmarks. CJ needs to take the data and create jobs from it.
git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/starexec/dev/starexec@5762 bd178f5f-7bea-4014-844f-495ca19b8805

+++ src/manage/Jobject.java
@@ -66,7 +66,7 @@ public class Jobject {
 	 * @param Solver ID
 	 * @return The SolverLink just created.
 	 */
-	public SolverLink addSolver(Long sid) throws Exception {
+	public SolverLink addSolver(int sid) throws Exception {
 		SolverLink lnk = new SolverLink(sid);
 		addLink(lnk);
 		return lnk;
